isadmin check on delete, edit, view and create

+++ routes/web.php
@@ -20,13 +20,11 @@ Auth::routes();
 
 Route::get('/', [CategoryController::class, 'index'])->name('home');
 
-Route::delete('/categories/delete/{id}', [CategoryController::class, 'destroy'])->name('categories.delete');
-
-Route::get('/categories/edit/{id}', [CategoryController::class, 'edit'])->name('categories.edit');
-Route::patch('/categories/update/{id}', [CategoryController::class, 'update'])->name('categories.update');
-
-Route::get('/categories/create', [CategoryController::class, 'create'])->name('categories.create');
-Route::post('/categories/store', [CategoryController::class, 'store'])->name('categories.store');
+Route::delete('/categories/delete/{id}', [CategoryController::class, 'destroy'])->name('categories.delete')->middleware('auth','isAdmin');
+Route::get('/categories/edit/{id}', [CategoryController::class, 'edit'])->name('categories.edit')->middleware('auth','isAdmin');
+Route::patch('/categories/update/{id}', [CategoryController::class, 'update'])->name('categories.update')->middleware('auth','isAdmin');
+Route::get('/categories/create', [CategoryController::class, 'create'])->name('categories.create')->middleware('auth','isAdmin');
+Route::post('/categories/store', [CategoryController::class, 'store'])->name('categories.store')->middleware('auth','isAdmin');
 
 Route::get('/categories/show/{id}', [CategoryController::class, 'show'])->name('categories.show');
